+++ src/EigenValueSolvers.jl
@@ -2,7 +2,7 @@
     EigenValueSolvers
 
 A thin, uniform interface around the eigenvalue solvers of `LinearAlgebra`, `Arpack.jl`,
-`ArnoldiMethod.jl` and `KrylovKit.jl`.
+`ArnoldiMethod.jl`, `KrylovKit.jl`, `IterativeSolvers.jl` and `GenericArpack.jl`.
 
 This package is heavily inspired by
 [`BifurcationKit.jl`](https://github.com/bifurcationkit/BifurcationKit.jl/blob/master/src/EigSolver.jl),
@@ -10,12 +10,12 @@ which is released under the MIT "Expat" License.
 """
 module EigenValueSolvers
 
-using LinearAlgebra: LinearAlgebra, eigen
+using LinearAlgebra: LinearAlgebra, eigen, lu
 using ArgCheck: @argcheck
 
 export AbstractEigenSolver, AbstractDirectEigenSolver, AbstractIterativeEigenSolver, AbstractMFEigenSolver
-export DefaultEig, EigArpack, EigKrylovKit, EigArnoldiMethod
-export gev, geteigenvector, getsolver
+export EigDefault, EigArpack, EigKrylovKit, EigArnoldiMethod, EigLOBPCG, EigGenericArpack
+export gev, geteigenvector, getsolver, supportedtargets
 
 include("interface.jl")
 include("solvers.jl")